requirements.txt
alembic==1.8.1
amqp==5.0.5
aniso8601==8.0.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
prompt-toolkit==3.0.16
psycopg==3.1.3
PyJWT==2.0.1
pymongo==3.11.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1